Move runStartCmdConsole back to runtime.go
